test: new e2e timezone tests
